more on the new resistor part
git-svn-id: http://fritzing.googlecode.com/svn/trunk@3512 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/modelbase.cpp
@@ -169,7 +169,7 @@ bool ModelBase::loadInstances(QDomElement & instances, QList<ModelPart *> & mode
 			if (nodeName.compare("originalModelIndex") == 0) continue;
 			if (nodeName.compare("path") == 0) continue;
 
-			modelPart->setProp(nodeName.toUtf8().constData(), node.nodeValue().toDouble());
+			modelPart->setProp(nodeName.toUtf8().constData(), node.nodeValue());
 		}
 
    		instance = instance.nextSiblingElement("instance");
